feat(views): move pager from list partial to index
We use a list partial in our index view. That allows to only render the
list partial using AJAX in use cases like filtering. In that case the
pagination links should also be updated. To enable this behaviour we
move the pager from the index to the list.

This patch moves the pagination helper from list to index view.